Refactor mcts.State to be a subclass of GameState called MCTSGameState. Add a constructor from GameState to MCTSGameState.
